+++ sdk/core/azure-core/test/ut/policy.cpp
@@ -62,7 +62,7 @@ struct TestContextTreeIntegrity : public Azure::Core::Http::Policies::HttpPolicy
     EXPECT_TRUE(ctx.HasKey("TheKey"));
     if (ctx.HasKey("TheKey"))
     {
-      auto value = ctx.Get<std::string>("TheKey");
+      auto value = ctx.GetValue<std::string>("TheKey");
       EXPECT_EQ("TheValue", value);
     }
     return nextHttpPolicy.Send(request, ctx);
@@ -223,6 +223,6 @@ TEST(Policy, RetryPolicyKeepContext)
   HttpPipeline pipeline(policies);
   Request request(HttpMethod::Get, Url("url"));
   auto withValueContext
-      = Context::GetApplicationContext().WithValue("TheKey", std::string("TheValue"));
+      = Context::GetApplicationContext().CreateWithValue("TheKey", std::string("TheValue"));
   pipeline.Send(request, withValueContext);
 }