+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/key_client_delete_test_live.cpp
@@ -59,7 +59,7 @@ TEST_F(KeyVaultClientTest, DeleteKey)
     // Setting 3 min as timeout just because I like number 3. We just want to prevent test running
     // for so long if something happens and no exception is thrown (paranoid scenario)
     auto duration = std::chrono::system_clock::now() + std::chrono::minutes(3);
-    auto cancelToken = Azure::Core::Context::GetApplicationContext().WithDeadline(duration);
+    auto cancelToken = Azure::Core::Context::GetApplicationContext().CreateWithExpiration(duration);
 
     auto keyResponseLRO = keyClient.StartDeleteKey(keyName);
     auto expectedStatusToken = m_keyVaultUrl
@@ -147,7 +147,7 @@ TEST_F(KeyVaultClientTest, DoubleDelete)
   }
   {
     auto duration = std::chrono::system_clock::now() + std::chrono::minutes(3);
-    auto cancelToken = Azure::Core::Context::GetApplicationContext().WithDeadline(duration);
+    auto cancelToken = Azure::Core::Context::GetApplicationContext().CreateWithExpiration(duration);
     auto keyResponseLRO = keyClient.StartDeleteKey(keyName);
     auto keyResponse = keyResponseLRO.PollUntilDone(std::chrono::milliseconds(1000), cancelToken);
   }
@@ -221,7 +221,7 @@ TEST_F(KeyVaultClientTest, CreateDeletedKey)
   }
   {
     auto duration = std::chrono::system_clock::now() + std::chrono::minutes(3);
-    auto cancelToken = Azure::Core::Context::GetApplicationContext().WithDeadline(duration);
+    auto cancelToken = Azure::Core::Context::GetApplicationContext().CreateWithExpiration(duration);
     auto keyResponseLRO = keyClient.StartDeleteKey(keyName);
     auto keyResponse = keyResponseLRO.PollUntilDone(std::chrono::milliseconds(1000), cancelToken);
   }
@@ -302,7 +302,7 @@ TEST_F(KeyVaultClientTest, GetDeletedKey)
   {
     // Wait until key is deleted
     auto duration = std::chrono::system_clock::now() + std::chrono::minutes(3);
-    auto cancelToken = Azure::Core::Context::GetApplicationContext().WithDeadline(duration);
+    auto cancelToken = Azure::Core::Context::GetApplicationContext().CreateWithExpiration(duration);
 
     auto keyResponseLRO = keyClient.StartDeleteKey(keyName);
     auto expectedStatusToken = m_keyVaultUrl