+++ sdk/storage/azure-storage-common/src/storage_switch_to_secondary_policy.cpp
@@ -13,7 +13,7 @@ namespace Azure { namespace Storage { namespace _internal {
     std::shared_ptr<bool> replicaStatus;
     if (ctx.HasKey(SecondaryHostReplicaStatusKey))
     {
-      replicaStatus = ctx.Get<std::shared_ptr<bool>>(SecondaryHostReplicaStatusKey);
+      replicaStatus = ctx.GetValue<std::shared_ptr<bool>>(SecondaryHostReplicaStatusKey);
     }
 
     bool considerSecondary = (request.GetMethod() == Azure::Core::Http::HttpMethod::Get